Return connection errors for producer requests
Producer now internally tracks NSQ connection errors, so that while it
is not connected to NSQ, the errors are returned for message publication
requests. Before this change, Producer would wait indefinitely to
publish a message until the connection was established.

This partially reverts some changes in commit 1d78b8b (but still leaves
out the old force-connect option).