Make connection errors for producer optional (#47)
The new ability for producers to return connection errors introduced in
PR #46 is made optional. The default behavior for producers reverts to
what it was: blocking the submission of messages on connection failure.
When the new `FailOnConErr` flag in `ProducerConfig` is set to `true`,
then the new error support is active instead.